python3/python_project/project/requirements.txt
# pip install packages
openpyxl
pyyaml
loguru
configparser

